+++ src/directives/ctr-completer.ts
@@ -14,33 +14,35 @@ export interface CompleterDropdown {
     selectCurrent(): void;
     nextRow(): void;
     prevRow(): void;
+    highlightRow(index: number | null): void;
 }
 
 @Directive({
     selector: "[ctrCompleter]",
 })
 export class CtrCompleter {
-    @Output() public selected = new EventEmitter<CompleterItem>();
-    @Output() public highlighted = new EventEmitter<CompleterItem>();
+    @Output() public selected = new EventEmitter<CompleterItem | null>();
+    @Output() public highlighted = new EventEmitter<CompleterItem | null>();
     @Output() public opened = new EventEmitter<boolean>();
+    @Output() public dataSourceChange = new EventEmitter<void>();
 
     private list: CompleterList;
-    private dropdown: CompleterDropdown;
+    private dropdown: CompleterDropdown | null;
     private _hasHighlighted = false;
     private _hasSelected = false;
     private _cancelBlur = false;
     private _isOpen = false;
-    private _autoHighlightIndex: number;
+    private _autoHighlightIndex: number | null;
 
     public registerList(list: CompleterList) {
         this.list = list;
     }
 
-    public registerDropdown(dropdown: CompleterDropdown) {
+    public registerDropdown(dropdown: CompleterDropdown | null) {
         this.dropdown = dropdown;
     }
 
-    public onHighlighted(item: CompleterItem) {
+    public onHighlighted(item: CompleterItem | null) {
         this.highlighted.emit(item);
         this._hasHighlighted = !!item;
     }
@@ -55,6 +57,14 @@ export class CtrCompleter {
         }
     }
 
+    public onDataSourceChange() {
+        if (this.hasSelected) {
+            this.selected.emit(null);
+            this._hasSelected = false;
+        }
+        this.dataSourceChange.emit();
+    }
+
     public search(term: string) {
         if (this._hasSelected) {
             this.selected.emit(null);
@@ -66,14 +76,16 @@ export class CtrCompleter {
     }
 
     public clear() {
+        this._hasHighlighted = false;
+        this.isOpen = false;
+
         if (this.dropdown) {
             this.dropdown.clear();
         }
+
         if (this.list) {
             this.list.clear();
         }
-        this._hasHighlighted = false;
-        this.isOpen = false;
     }
 
     public selectCurrent() {
@@ -129,8 +141,11 @@ export class CtrCompleter {
         return this._autoHighlightIndex;
     }
 
-    public set autoHighlightIndex(index: number) {
+    public set autoHighlightIndex(index: number | null) {
         this._autoHighlightIndex = index;
+        if (this.dropdown) {
+            this.dropdown.highlightRow(this._autoHighlightIndex);
+        }
     }
 
     public get hasSelected() {