+++ src/directives/ctr-list.ts
@@ -1,8 +1,9 @@
 import "rxjs/add/observable/timer";
-import { ChangeDetectorRef, Directive, Host, Input, OnInit, TemplateRef, ViewContainerRef } from "@angular/core";
+import { ChangeDetectorRef, Directive, EmbeddedViewRef, Host, Input, OnInit, TemplateRef, ViewContainerRef } from "@angular/core";
 import { Observable } from "rxjs/Observable";
 import { Subscription } from "rxjs/Subscription";
-
+import "rxjs/add/operator/retryWhen";
+import "rxjs/add/operator/do";
 
 import { CtrCompleter, CompleterList } from "./ctr-completer";
 import { CompleterData } from "../services/completer-data";
@@ -12,7 +13,7 @@ import { MIN_SEARCH_LENGTH, PAUSE, CLEAR_TIMEOUT, isNil } from "../globals";
 
 export class CtrListContext {
     constructor(
-        public results: CompleterItem[],
+        public results: CompleterItem[] | null,
         public searching: boolean,
         public searchInitialized: boolean,
         public isOpen: boolean
@@ -27,15 +28,17 @@ export class CtrList implements OnInit, CompleterList {
     @Input() public ctrListPause = PAUSE;
     @Input() public ctrListAutoMatch = false;
     @Input() public ctrListAutoHighlight = false;
+    @Input() public ctrListDisplaySearching = true;
 
     private _dataService: CompleterData;
     // private results: CompleterItem[] = [];
-    private term: string = null;
+    private term: string | null = null;
     // private searching = false;
-    private searchTimer: Subscription = null;
-    private clearTimer: Subscription = null;
+    private searchTimer: Subscription | null = null;
+    private clearTimer: Subscription | null = null;
     private ctx = new CtrListContext([], false, false, false);
     private _initialValue: any = null;
+    private viewRef: EmbeddedViewRef<CtrListContext> | null = null;
 
     constructor(
         @Host() private completer: CtrCompleter,
@@ -45,7 +48,7 @@ export class CtrList implements OnInit, CompleterList {
 
     public ngOnInit() {
         this.completer.registerList(this);
-        this.viewContainer.createEmbeddedView(
+        this.viewRef = this.viewContainer.createEmbeddedView(
             this.templateRef,
             new CtrListContext([], false, false, false)
         );
@@ -54,35 +57,14 @@ export class CtrList implements OnInit, CompleterList {
     @Input("ctrList")
     public set dataService(newService: CompleterData) {
         this._dataService = newService;
-        if (this._dataService) {
-            this._dataService
-                .catch(err => this.handleError(err))
-                .subscribe(results => {
-                    this.ctx.searchInitialized = true;
-                    this.ctx.searching = false;
-                    this.ctx.results = results;
-                    if (this.ctrListAutoMatch && results.length === 1 && results[0].title && !isNil(this.term) &&
-                        results[0].title.toLocaleLowerCase() === this.term.toLocaleLowerCase()) {
-                        // Do automatch
-                        this.completer.onSelected(results[0]);
-                    }
-                    if (this._initialValue) {
-                        this.initialValue = this._initialValue;
-                        this._initialValue = null;
-                    }
-                    if (this.ctrListAutoHighlight) {
-                        this.completer.autoHighlightIndex = this.getBestMatchIndex();
-                    }
-                    this.refreshTemplate();
-                });
-        }
+        this.dataServiceSubscribe();
     }
 
     @Input("ctrListInitialValue")
     public set initialValue(value: any) {
         if (this._dataService && typeof this._dataService.convertToItem === "function") {
             setTimeout(() => {
-                const initialItem = this._dataService.convertToItem(value);
+                const initialItem = this._dataService.convertToItem!(value);
                 if (initialItem) {
                     this.completer.onSelected(initialItem, false);
                 }
@@ -99,7 +81,9 @@ export class CtrList implements OnInit, CompleterList {
                 this.searchTimer = null;
             }
             if (!this.ctx.searching) {
-                this.ctx.results = [];
+                if (this.ctrListDisplaySearching) {
+                    this.ctx.results = [];
+                }
                 this.ctx.searching = true;
                 this.ctx.searchInitialized = true;
                 this.refreshTemplate();
@@ -112,6 +96,7 @@ export class CtrList implements OnInit, CompleterList {
             });
         } else if (!isNil(term) && term.length < this.ctrListMinSearchLength) {
             this.clear();
+            this.term = "";
         }
     }
 
@@ -145,6 +130,7 @@ export class CtrList implements OnInit, CompleterList {
         }
 
         this.viewContainer.clear();
+        this.viewRef = null;
     }
 
     private searchTimerComplete(term: string) {
@@ -157,51 +143,86 @@ export class CtrList implements OnInit, CompleterList {
         this._dataService.search(term);
     }
 
-    private handleError(error: any) {
-        this.ctx.searching = false;
-        let errMsg: string = "search error";
-        if (error) {
-            errMsg = (error.message) ? error.message :
-                error.status ? `${error.status} - ${error.statusText}` : "Server error";
-        }
-
-        if (console && console.error) {
-            console.error(errMsg); // log to console
-        }
-        this.refreshTemplate();
-
-        return Observable.throw(errMsg);
-    }
-
     private refreshTemplate() {
-        // Recreate the template
-        this.viewContainer.clear();
-        if (this.ctx.results && this.ctx.isOpen) {
-            this.viewContainer.createEmbeddedView(
+        // create the template if it doesn't exist
+        if (!this.viewRef) {
+            this.viewRef = this.viewContainer.createEmbeddedView(
                 this.templateRef,
                 this.ctx
             );
+        } else if (!this.viewRef.destroyed) {
+            // refresh the template
+            this.viewRef!.context.isOpen = this.ctx.isOpen;
+            this.viewRef!.context.results = this.ctx.results;
+            this.viewRef!.context.searching = this.ctx.searching;
+            this.viewRef!.context.searchInitialized = this.ctx.searchInitialized;
+            this.viewRef.detectChanges();
         }
         this.cd.markForCheck();
     }
 
     private getBestMatchIndex() {
-        if (!this.ctx.results) {
+        if (!this.ctx.results || !this.term) {
             return null;
         }
 
         // First try to find the exact term
-        let bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase() === this.term.toLocaleLowerCase());
+        let bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase() === this.term!.toLocaleLowerCase());
         // If not try to find the first item that starts with the term
         if (bestMatch < 0) {
-            bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase().startsWith(this.term.toLocaleLowerCase()));
+            bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase().startsWith(this.term!.toLocaleLowerCase()));
         }
         // If not try to find the first item that includes the term
         if (bestMatch < 0) {
-            bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase().includes(this.term.toLocaleLowerCase()));
+            bestMatch = this.ctx.results.findIndex(item => item.title.toLowerCase().includes(this.term!.toLocaleLowerCase()));
         }
 
         return bestMatch < 0 ? null : bestMatch;
+    }
+
+    private dataServiceSubscribe() {
+        if (this._dataService) {
+            this._dataService
+                .catch(err => {
+                    console.error(err);
+                    console.error("Unexpected error in dataService: errors should be handled by the dataService Observable");
+                    return [];
+                })
+                .subscribe(results => {
+                    this.ctx.searchInitialized = true;
+                    this.ctx.searching = false;
+                    this.ctx.results = results;
+
+                    if (this.ctrListAutoMatch && results && results.length === 1 && results[0].title && !isNil(this.term) &&
+                        results[0].title.toLocaleLowerCase() === this.term!.toLocaleLowerCase()) {
+                        // Do automatch
+                        this.completer.onSelected(results[0]);
+                        return;
+                    }
+
+                    if (this._initialValue) {
+                        this.initialValue = this._initialValue;
+                        this._initialValue = null;
+                    }
+
+                    this.refreshTemplate();
+
+                    if (this.ctrListAutoHighlight) {
+                        this.completer.autoHighlightIndex = this.getBestMatchIndex();
+                    }
+                });
+
+            if (this._dataService.dataSourceChange) {
+                this._dataService.dataSourceChange.subscribe(() => {
+                    this.term = null;
+                    this.ctx.searchInitialized = false;
+                    this.ctx.searching = false;
+                    this.ctx.results = [];
+                    this.refreshTemplate();
+                    this.completer.onDataSourceChange();
+                });
+            }
+        }
     }
 
 }